--- ATM_Project/src/main/java/com/atm/demo/TestController.java
@@ -1,13 +1,0 @@
-package com.atm.demo;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TestController {
-
-    @GetMapping("/hello")
-    public String hello(){
-        return "Atm works fine!";
-    }
-}
