--- ATM_Project/src/main/java/com/atm/demo/model/User.java
@@ -1,30 +1,0 @@
-package com.atm.demo.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "users")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false,unique = true)
-    private String cardNumber;
-
-    @Column(nullable = false)
-    private String pin;
-
-    @Column(nullable = false)
-    private Double balance;
-
-    @Column(nullable = false)
-    private String role;
-}
