--- ATM_Project/src/main/java/com/atm/demo/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.atm.demo.repository;
-
-import com.atm.demo.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByCardNumber(String cardNumber);
-}
